# 20% speed up of the GPU NeighborsFinder