20% speed up of the GPU NeighborsFinder via manual loop unrolling.

before:
AMD: 13,692 us NVIDIA: 5,169 us

after:
AMD: 11,406 us NVIDIA: 4,129 us