@davidrohr 2020-03-24
@sgorbuno : OpenCL 1.2 doesn't know constexpr, you can use CONSTEXPR.
For the #pragma unroll, I am not sure whether opencl will work with a variable (even if const) as unroll factor.
One point I was thinking about, is to have a for loop with 2 iterations and pragma unroll 2 around your splitting of the loop. First iteration could do the shared memory part, 2nd one the scratch memory part. If the compiler unrolls correctly, that should compile to the same code you have right now. But not sure whether the HIP compiler will understand it and how many lines of code we can save this way.