@sgorbuno 2020-03-24
> For the #pragma unroll, I am not sure whether opencl will work with a variable (even if const) as unroll factor.

nvcc doesn't work with a macro as an unroll factor. That is why I introduced constexpr ...