@davidrohr 2020-03-24
you can use GPUCA_UNROLL

the syntax is GPUCA_UNROLL(CUDA, HIP), where CUDA and HIP can be:
- empty: no unroll
- U() : becomes #pragma unroll
- U(n) : becomes #pragma unroll n

I also have to add OpenCL and CPU here, but the time being, this should
be sufficient for you I guess, in combination with CONSTEXPR

On 24.03.20 12:37, Sergey Gorbunov wrote:
>
>     For the #pragma unroll, I am not sure whether opencl will work
>     with a variable (even if const) as unroll factor.
>
> nvcc doesn't work with a macro as an unroll factor. That is why I
> introduced constexpr ...
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/AliceO2Group/AliceO2/pull/3213#issuecomment-603189005>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACFVM5TFKMIV6A77GZRHJH3RJCLRHANCNFSM4LSJNLRA>.
>