@davidrohr 2020-03-25
@sgorbuno : I think you can try to use the GPUCA_UNROLL macro, to avoid the OpenCL problems, as discussed via e-mail.